fixed modules and documentation issues #24128
