+++ framework/doc/content/source/actions/AddConvergenceAction.md
@@ -1,0 +1,8 @@
+# AddConvergenceAction
+
+!syntax description /Convergence/AddConvergenceAction
+
+More information about `Convergence` may be found on the
+[Convergence syntax documentation](syntax/Convergence/index.md).
+
+!syntax parameters /Convergence/AddConvergenceAction
