+++ framework/include/convergence/Convergence.h
@@ -1,0 +1,36 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "MooseObject.h"
+#include "SetupInterface.h"
+#include "PerfGraphInterface.h"
+
+class Convergence : public MooseObject, public SetupInterface, public PerfGraphInterface
+{
+public:
+  static InputParameters validParams();
+
+  enum class MooseConvergenceStatus
+  {
+    ITERATING = 0,
+    CONVERGED = 2,
+    DIVERGED = -2
+  };
+
+  Convergence(const InputParameters & parameters);
+
+  virtual void initialSetup() override{};
+
+  virtual MooseConvergenceStatus checkConvergence(int it, Real xnorm, Real snorm, Real fnorm) = 0;
+
+protected:
+  PerfID _perf_check_convergence;
+};
