+++ framework/include/problems/FEProblemBase.h
@@ -85,7 +85,7 @@ class LineSearch;
 class UserObject;
 class AutomaticMortarGeneration;
 class VectorPostprocessor;
-class Function;
+class Convergence;
 class MooseAppCoordTransform;
 class MortarUserObject;
 
@@ -222,52 +222,6 @@ public:
   couplingEntries(const THREAD_ID tid, const unsigned int nl_sys_num);
   std::vector<std::pair<MooseVariableFEBase *, MooseVariableFEBase *>> &
   nonlocalCouplingEntries(const THREAD_ID tid, const unsigned int nl_sys_num);
-
-  /**
-   * Check for convergence of the nonlinear solution
-   * @param msg Error message that gets sent back to the solver
-   * @param it Iteration counter
-   * @param xnorm Norm of the solution vector
-   * @param snorm Norm of the change in the solution vector
-   * @param fnorm Norm of the residual vector
-   * @param rtol Relative residual convergence tolerance
-   * @param divtol Relative residual divergence tolerance
-   * @param stol Solution change convergence tolerance
-   * @param abstol Absolute residual convergence tolerance
-   * @param nfuncs Number of function evaluations
-   * @param max_funcs Maximum Number of function evaluations
-   * @param div_threshold Maximum value of residual before triggering divergence check
-   */
-  virtual MooseNonlinearConvergenceReason checkNonlinearConvergence(std::string & msg,
-                                                                    const PetscInt it,
-                                                                    const Real xnorm,
-                                                                    const Real snorm,
-                                                                    const Real fnorm,
-                                                                    const Real rtol,
-                                                                    const Real divtol,
-                                                                    const Real stol,
-                                                                    const Real abstol,
-                                                                    const PetscInt nfuncs,
-                                                                    const PetscInt max_funcs,
-                                                                    const Real div_threshold);
-
-  /// Perform steps required before checking nonlinear convergence
-  virtual void nonlinearConvergenceSetup() {}
-
-  /**
-   * Check the relative convergence of the nonlinear solution
-   * @param fnorm          Norm of the residual vector
-   * @param the_residual   The residual to check
-   * @param rtol           Relative tolerance
-   * @param abstol         Absolute tolerance
-   * @return               Bool signifying convergence
-   */
-  virtual bool checkRelativeConvergence(const PetscInt it,
-                                        const Real fnorm,
-                                        const Real the_residual,
-                                        const Real rtol,
-                                        const Real abstol,
-                                        std::ostringstream & oss);
 
   virtual bool hasVariable(const std::string & var_name) const override;
   using SubProblem::getVariable;
@@ -675,6 +629,13 @@ public:
   /// Get a MeshDivision
   MeshDivision & getMeshDivision(const std::string & name, const THREAD_ID tid = 0) const;
 
+  virtual void
+  addConvergence(const std::string & type, const std::string & name, InputParameters & parameters);
+  virtual Convergence & getConvergence(const std::string & name, const THREAD_ID tid = 0) const;
+  virtual const std::vector<std::shared_ptr<Convergence>> &
+  getConvergenceObjects(const THREAD_ID tid = 0) const;
+  virtual bool hasConvergence(const std::string & name, const THREAD_ID tid = 0) const;
+  virtual void addDefaultConvergence();
   /**
    * add a MOOSE line search
    */
@@ -2199,6 +2160,8 @@ public:
     _n_max_nl_pingpong = n_max_nl_pingpong;
   }
 
+  unsigned int getMaxNLPingPong() { return _n_max_nl_pingpong; }
+
   /// method setting the minimum number of nonlinear iterations before performing divergence checks
   void setNonlinearForcedIterations(const unsigned int nl_forced_its)
   {
@@ -2214,6 +2177,19 @@ public:
     _nl_abs_div_tol = nl_abs_div_tol;
   }
 
+  /// method returning the absolute divergence tolerance
+  Real getNonlinearAbsoluteDivergenceTolerance() const { return _nl_abs_div_tol; };
+
+  /**
+   * Sets the nonlinear convergence object name if there is one
+   */
+  void setNonlinearConvergenceName(const ConvergenceName & convergence)
+  {
+    _nonlinear_convergence_name = convergence;
+    _set_nonlinear_convergence_name = true;
+  }
+
+  ConvergenceName getNonlinearConvergenceName() const { return _nonlinear_convergence_name; }
   /**
    * Setter for whether we're computing the scaling jacobian
    */
@@ -2284,6 +2260,10 @@ public:
    */
   void setFailNextNonlinearConvergenceCheck() { _fail_next_nonlinear_convergence_check = true; }
 
+  /**
+   * Do not skip further residual evaluations and fail the next nonlinear convergence check
+   */
+  void resetFailNextNonlinearConvergenceCheck() { _fail_next_nonlinear_convergence_check = false; }
   /*
    * Set the status of loop order of execution printing
    * @param print_exec set of execution flags to print on
@@ -2387,6 +2367,9 @@ private:
 protected:
   bool _initialized;
 
+  /// Nonlinear convergence name
+  ConvergenceName _nonlinear_convergence_name;
+
   std::set<TagID> _fe_vector_tags;
 
   std::set<TagID> _fe_matrix_tags;
@@ -2407,7 +2390,11 @@ protected:
   Real & _dt;
   Real & _dt_old;
 
-  /// maximum numbver
+  /// Flag that the nonlinear convergence name has been set
+  bool _set_nonlinear_convergence_name;
+  bool _set_reference_convergence_name = false;
+
+  /// maximum number
   unsigned int _n_nl_pingpong = 0;
   unsigned int _n_max_nl_pingpong = std::numeric_limits<unsigned int>::max();
 
@@ -2482,6 +2469,9 @@ protected:
 
   /// functions
   MooseObjectWarehouse<Function> _functions;
+
+  /// convergence warehouse
+  MooseObjectWarehouse<Convergence> _convergences;
 
   /// nonlocal kernels
   MooseObjectWarehouse<KernelBase> _nonlocal_kernels;
