+++ framework/include/utils/MooseTypes.h
@@ -1143,6 +1143,9 @@ DerivativeStringClass(ParsedFunctionExpression);
 /// System name support of multiple nonlinear systems on the same mesh
 DerivativeStringClass(NonlinearSystemName);
 
+/// Name of a Convergence object
+DerivativeStringClass(ConvergenceName);
+
 /// System name support of multiple linear systems on the same mesh
 DerivativeStringClass(LinearSystemName);
