+++ framework/src/base/Moose.C
@@ -188,6 +188,8 @@ addActionTypes(Syntax & syntax)
   registerMooseObjectTask("add_control",                  Control,                   false);
   registerMooseObjectTask("add_partitioner",              MoosePartitioner,          false);
 
+  registerMooseObjectTask("add_convergence",              Convergence,            false);
+
   // clang-format on
 
   registerTask("dynamic_object_registration", false);
@@ -324,6 +326,7 @@ addActionTypes(Syntax & syntax)
                            "(add_mortar_variable)"
                            "(setup_variable_complete)"
                            "(setup_quadrature)"
+                           "(add_convergence)"
                            "(add_periodic_bc)"
                            "(add_user_object)"
                            "(add_distribution)"
@@ -472,6 +475,8 @@ associateSyntaxInner(Syntax & syntax, ActionFactory & /*action_factory*/)
 
   registerSyntax("AddMeshDivisionAction", "MeshDivisions/*");
   syntax.registerSyntaxType("MeshDivisions/*", "MeshDivisionName");
+  registerSyntax("AddConvergenceAction", "Convergence/*");
+  syntax.registerSyntaxType("Convergence/*", "ConvergenceName");
 
   registerSyntax("GlobalParamsAction", "GlobalParams");
