+++ framework/src/executioners/FEProblemSolve.C
@@ -11,6 +11,8 @@
 
 #include "FEProblem.h"
 #include "NonlinearSystemBase.h"
+#include "ResidualConvergence.h"
+#include "ReferenceResidualConvergence.h"
 
 std::set<std::string> const FEProblemSolve::_moose_line_searches = {"contact", "project"};
 
@@ -21,9 +23,40 @@ FEProblemSolve::mooseLineSearches()
 }
 
 InputParameters
+FEProblemSolve::residualConvergenceParams()
+{
+  InputParameters params = emptyInputParameters();
+  params.addParam<Real>("l_tol", 1.0e-5, "Linear Relative Tolerance");
+  params.addParam<Real>("l_abs_tol", 1.0e-50, "Linear Absolute Tolerance");
+  params.addParam<unsigned int>("l_max_its", 10000, "Max Linear Iterations");
+  params.addParam<unsigned int>("nl_max_its", 50, "Max Nonlinear Iterations");
+  params.addParam<unsigned int>("nl_forced_its", 0, "The Number of Forced Nonlinear Iterations");
+  params.addParam<unsigned int>("nl_max_funcs", 10000, "Max Nonlinear solver function evaluations");
+  params.addParam<Real>("nl_abs_tol", 1.0e-50, "Nonlinear Absolute Tolerance");
+  params.addParam<Real>("nl_rel_tol", 1.0e-8, "Nonlinear Relative Tolerance");
+  params.addParam<Real>(
+      "nl_div_tol",
+      1.0e10,
+      "Nonlinear Relative Divergence Tolerance. A negative value disables this check.");
+  params.addParam<Real>(
+      "nl_abs_div_tol",
+      1.0e50,
+      "Nonlinear Absolute Divergence Tolerance. A negative value disables this check.");
+  params.addParam<Real>("nl_abs_step_tol", 0., "Nonlinear Absolute step Tolerance");
+  params.addParam<Real>("nl_rel_step_tol", 0., "Nonlinear Relative step Tolerance");
+  params.addParam<unsigned int>(
+      "n_max_nonlinear_pingpong",
+      100,
+      "The maximum number of times the nonlinear residual can ping pong "
+      "before requesting halting the current evaluation and requesting timestep cut");
+  return params;
+}
+
+InputParameters
 FEProblemSolve::validParams()
 {
   InputParameters params = emptyInputParameters();
+  params += FEProblemSolve::residualConvergenceParams();
 
   params.addParam<std::vector<std::string>>("splitting",
                                             {},
@@ -58,29 +91,9 @@ FEProblemSolve::validParams()
                         "be looser than the standard linear tolerance");
 
   params += Moose::PetscSupport::getPetscValidParams();
-  params.addParam<Real>("l_tol", 1.0e-5, "Linear Relative Tolerance");
-  params.addParam<Real>("l_abs_tol", 1.0e-50, "Linear Absolute Tolerance");
-  params.addParam<unsigned int>("l_max_its", 10000, "Max Linear Iterations");
-  params.addParam<unsigned int>("nl_max_its", 50, "Max Nonlinear Iterations");
-  params.addParam<unsigned int>("nl_forced_its", 0, "The Number of Forced Nonlinear Iterations");
-  params.addParam<unsigned int>("nl_max_funcs", 10000, "Max Nonlinear solver function evaluations");
-  params.addParam<Real>("nl_abs_tol", 1.0e-50, "Nonlinear Absolute Tolerance");
-  params.addParam<Real>("nl_rel_tol", 1.0e-8, "Nonlinear Relative Tolerance");
-  params.addParam<Real>(
-      "nl_div_tol",
-      1.0e10,
-      "Nonlinear Relative Divergence Tolerance. A negative value disables this check.");
-  params.addParam<Real>(
-      "nl_abs_div_tol",
-      1.0e50,
-      "Nonlinear Absolute Divergence Tolerance. A negative value disables this check.");
-  params.addParam<Real>("nl_abs_step_tol", 0., "Nonlinear Absolute step Tolerance");
-  params.addParam<Real>("nl_rel_step_tol", 0., "Nonlinear Relative step Tolerance");
-  params.addParam<unsigned int>(
-      "n_max_nonlinear_pingpong",
-      100,
-      "The maximum number of times the nonlinear residual can ping pong "
-      "before requesting halting the current evaluation and requesting timestep cut");
+  params.addParam<ConvergenceName>(
+      "nonlinear_convergence",
+      "Name of Convergence object to use to assess convergence of the nonlinear solve");
   params.addParam<bool>(
       "snesmf_reuse_base",
       true,
@@ -225,6 +238,16 @@ FEProblemSolve::FEProblemSolve(Executioner & ex)
   _problem.setNonlinearForcedIterations(getParam<unsigned int>("nl_forced_its"));
 
   _problem.setNonlinearAbsoluteDivergenceTolerance(getParam<Real>("nl_abs_div_tol"));
+
+  if (isParamValid("nonlinear_convergence"))
+  {
+    _problem.setNonlinearConvergenceName(getParam<ConvergenceName>("nonlinear_convergence"));
+  }
+  else
+  {
+    _problem.setNonlinearConvergenceName("default_convergence");
+    _problem.addDefaultConvergence();
+  }
 
   _nl.setDecomposition(_splitting);
