+++ framework/src/interfaces/TaggingInterface.C
@@ -11,6 +11,7 @@
 #include "Conversion.h"
 #include "FEProblem.h"
 #include "Assembly.h"
+#include "ReferenceResidualConvergence.h"
 #include "ReferenceResidualProblem.h"
 
 #include "libmesh/dense_vector.h"
@@ -74,7 +75,8 @@ TaggingInterface::TaggingInterface(const MooseObject * moose_object)
                    "' for Kernel '",
                    _moose_object.name(),
                    "' is not a residual vector tag");
-      _vector_tags.insert(vector_tag_id);
+      else
+        _vector_tags.insert(vector_tag_id);
     }
   }
 
@@ -91,7 +93,8 @@ TaggingInterface::TaggingInterface(const MooseObject * moose_object)
                  "' for Kernel '",
                  _moose_object.name(),
                  "' is not a residual vector tag");
-    _vector_tags.insert(vector_tag_id);
+    else
+      _vector_tags.insert(vector_tag_id);
   }
 
   // Add absolue value vector tags. These tags should be created in the System already, otherwise
@@ -127,25 +130,31 @@ TaggingInterface::TaggingInterface(const MooseObject * moose_object)
 
   const auto * const fe_problem =
       moose_object->parameters().getCheckedPointerParam<FEProblemBase *>("_fe_problem_base");
-  if (const auto * const ref_problem = dynamic_cast<const ReferenceResidualProblem *>(fe_problem))
+
+  for (auto conv_obj : fe_problem->getConvergenceObjects())
   {
-    const auto reference_tag = ref_problem->referenceVectorTagID({});
-    auto create_tags_split =
-        [reference_tag](const auto & tags, auto & non_ref_tags, auto & ref_tags)
+    if (conv_obj->name() == "ReferenceResidualProblem")
     {
-      for (const auto tag : tags)
-        if (tag == reference_tag)
-          ref_tags.insert(tag);
-        else
-          non_ref_tags.insert(tag);
-    };
-    create_tags_split(_vector_tags, _non_ref_vector_tags, _ref_vector_tags);
-    create_tags_split(_abs_vector_tags, _non_ref_abs_vector_tags, _ref_abs_vector_tags);
-  }
-  else
-  {
-    _non_ref_vector_tags = _vector_tags;
-    _non_ref_abs_vector_tags = _abs_vector_tags;
+      const auto * const ref_conv =
+          dynamic_cast<const ReferenceResidualConvergence *>(conv_obj.get());
+      const auto reference_tag = ref_conv->referenceVectorTagID({});
+      auto create_tags_split =
+          [reference_tag](const auto & tags, auto & non_ref_tags, auto & ref_tags)
+      {
+        for (const auto tag : tags)
+          if (tag == reference_tag)
+            ref_tags.insert(tag);
+          else
+            non_ref_tags.insert(tag);
+      };
+      create_tags_split(_vector_tags, _non_ref_vector_tags, _ref_vector_tags);
+      create_tags_split(_abs_vector_tags, _non_ref_abs_vector_tags, _ref_abs_vector_tags);
+    }
+    else
+    {
+      _non_ref_vector_tags = _vector_tags;
+      _non_ref_abs_vector_tags = _abs_vector_tags;
+    }
   }
 }
