+++ framework/src/parser/Builder.C
@@ -1075,6 +1075,7 @@ Builder::extractParams(const std::string & prefix, InputParameters & p)
         setscalar(SolverVariableName, string);
         setscalar(AuxVariableName, string);
         setscalar(FunctionName, string);
+        setscalar(ConvergenceName, string);
         setscalar(MeshDivisionName, string);
         setscalar(UserObjectName, string);
         setscalar(VectorPostprocessorName, string);
@@ -1149,6 +1150,7 @@ Builder::extractParams(const std::string & prefix, InputParameters & p)
         setvector(SolverVariableName, string);
         setvector(AuxVariableName, string);
         setvector(FunctionName, string);
+        setvector(ConvergenceName, string);
         setvector(MeshDivisionName, string);
         setvector(UserObjectName, string);
         setvector(IndicatorName, string);
@@ -1212,6 +1214,7 @@ Builder::extractParams(const std::string & prefix, InputParameters & p)
         setvectorvector(SolverVariableName);
         setvectorvector(AuxVariableName);
         setvectorvector(FunctionName);
+        setvectorvector(ConvergenceName);
         setvectorvector(UserObjectName);
         setvectorvector(IndicatorName);
         setvectorvector(MarkerName);
