+++ test/tests/convergence/reference_residual/tests
@@ -1,0 +1,11 @@
+[Tests]
+  design = 'ReferenceResidualConvergence.md'
+  issues = '#25931'
+  [reference_residual]
+    type = 'Exodiff'
+    input = 'reference_residual.i'
+    exodiff = 'reference_residual_out.e'
+    requirement = 'The system shall perform convergence checks by comparing individual '
+                  'variables to reference quantities of those variables.'
+  []
+[]
