+++ test/tests/convergence/residual_convergence/diffusion_convergence.i
@@ -1,0 +1,52 @@
+[Mesh]
+  type = GeneratedMesh
+  dim = 2
+  nx = 10
+  ny = 10
+[]
+
+[Variables]
+  [u]
+  []
+[]
+
+[Kernels]
+  [diff]
+    type = Diffusion
+    variable = u
+  []
+[]
+
+[BCs]
+  [left]
+    type = DirichletBC
+    variable = u
+    boundary = left
+    value = 0
+  []
+  [right]
+    type = DirichletBC
+    variable = u
+    boundary = right
+    value = 1
+  []
+[]
+
+[Convergence]
+  [res_conv]
+    type = ResidualConvergence
+    nl_abs_tol=1e-6
+  []
+[]
+
+[Executioner]
+  type = Steady
+  solve_type = 'PJFNK'
+  petsc_options_iname = '-pc_type'
+  petsc_options_value = 'hypre'
+  nonlinear_convergence = res_conv
+[]
+
+[Outputs]
+  exodus = true
+[]
