+++ test/tests/convergence/residual_convergence/tests
@@ -1,0 +1,10 @@
+[Tests]
+  design = 'ResidualConvergence.md'
+  issues = '#25931'
+  [nonlinear]
+    type = 'Exodiff'
+    input = 'diffusion_convergence.i'
+    exodiff = 'diffusion_convergence_out.e'
+    requirement = 'The system shall query the convergence status of the nonlinear solvers.'
+  []
+[]
